--- a/_examples/custom_ws_gobwas/transport.go
+++ b/_examples/custom_ws_gobwas/transport.go
@@ -40,6 +40,11 @@ func (t *customWebsocketTransport) Protocol() centrifuge.ProtocolType {
 	return t.protoType
 }
 
+// ProtocolVersion ...
+func (t *customWebsocketTransport) ProtocolVersion() centrifuge.ProtocolVersion {
+	return centrifuge.ProtocolVersion1
+}
+
 // Unidirectional returns whether transport is unidirectional.
 func (t *customWebsocketTransport) Unidirectional() bool {
 	return false
@@ -48,11 +53,6 @@ func (t *customWebsocketTransport) Unidirectional() bool {
 // DisabledPushFlags ...
 func (t *customWebsocketTransport) DisabledPushFlags() uint64 {
 	return centrifuge.PushFlagDisconnect
-}
-
-// Version of protocol.
-func (t *customWebsocketTransport) Version() centrifuge.ProtocolVersion {
-	return centrifuge.ProtocolVersion1
 }
 
 func (t *customWebsocketTransport) read() ([]byte, bool, error) {