--- a/_examples/go.mod
+++ b/_examples/go.mod
@@ -3,6 +3,7 @@ module github.com/centrifugal/centrifuge/_examples
 go 1.14
 
 replace github.com/centrifugal/centrifuge => ../
+replace github.com/centrifugal/protocol => ../../protocol
 
 require (
 	github.com/FZambia/tarantool v0.2.2