--- a/_examples/go.mod
+++ b/_examples/go.mod
@@ -3,12 +3,11 @@ module github.com/centrifugal/centrifuge/_examples
 go 1.14
 
 replace github.com/centrifugal/centrifuge => ../
-replace github.com/centrifugal/protocol => ../../protocol
 
 require (
 	github.com/FZambia/tarantool v0.2.2
 	github.com/centrifugal/centrifuge v0.8.2
-	github.com/centrifugal/protocol v0.7.7-0.20211217125855-f72f5e9399c6
+	github.com/centrifugal/protocol v0.7.7-0.20211220095654-956aec57883a
 	github.com/cristalhq/jwt/v3 v3.0.0
 	github.com/dchest/uniuri v0.0.0-20200228104902-7aecb25e1fe5
 	github.com/gin-contrib/sessions v0.0.3