--- a/_examples/redis_broker_presence/main.go
+++ b/_examples/redis_broker_presence/main.go
@@ -111,7 +111,7 @@ func main() {
 	broker, err := centrifuge.NewRedisBroker(node, centrifuge.RedisBrokerConfig{
 		// Use reasonably large expiration interval for stream meta key,
 		// much bigger than maximum HistoryLifetime value in Node config.
-		// This way stream meta data will expire, in some cases you may want
+		// This way stream metadata will expire, in some cases you may want
 		// to prevent its expiration setting this to zero value.
 		HistoryMetaTTL: 7 * 24 * time.Hour,