--- a/_examples/unidirectional_grpc/main.go
+++ b/_examples/unidirectional_grpc/main.go
@@ -190,6 +190,11 @@ func (t *grpcTransport) DisabledPushFlags() uint64 {
 	return 0
 }
 
+// Version of protocol.
+func (t *grpcTransport) Version() centrifuge.ProtocolVersion {
+	return centrifuge.ProtocolVersion1
+}
+
 func (t *grpcTransport) Write(message []byte) error {
 	return t.WriteMany(message)
 }