--- a/_examples/unidirectional_sse/main.go
+++ b/_examples/unidirectional_sse/main.go
@@ -268,6 +268,10 @@ func (t *eventsourceTransport) Protocol() centrifuge.ProtocolType {
 	return centrifuge.ProtocolTypeJSON
 }
 
+func (t *eventsourceTransport) ProtocolVersion() centrifuge.ProtocolVersion {
+	return centrifuge.ProtocolVersion1
+}
+
 // Unidirectional returns whether transport is unidirectional.
 func (t *eventsourceTransport) Unidirectional() bool {
 	return true
@@ -276,11 +280,6 @@ func (t *eventsourceTransport) Unidirectional() bool {
 // DisabledPushFlags ...
 func (t *eventsourceTransport) DisabledPushFlags() uint64 {
 	return 0
-}
-
-// Version of protocol.
-func (t *eventsourceTransport) Version() centrifuge.ProtocolVersion {
-	return centrifuge.ProtocolVersion1
 }
 
 func (t *eventsourceTransport) Write(message []byte) error {