--- a/_examples/ws_benchmarks/benchmark_gobwas/transport.go
+++ b/_examples/ws_benchmarks/benchmark_gobwas/transport.go
@@ -48,6 +48,11 @@ func (t *customWebsocketTransport) DisabledPushFlags() uint64 {
 	return centrifuge.PushFlagDisconnect
 }
 
+// Version of protocol.
+func (t *customWebsocketTransport) Version() centrifuge.ProtocolVersion {
+	return centrifuge.ProtocolVersion1
+}
+
 func (t *customWebsocketTransport) read() ([]byte, bool, error) {
 	t.mu.Lock()
 	defer t.mu.Unlock()