--- a/_examples/ws_benchmarks/benchmark_gorilla/main.go
+++ b/_examples/ws_benchmarks/benchmark_gorilla/main.go
@@ -95,8 +95,15 @@ func main() {
 		panic(err)
 	}
 
+	protocolVersion := centrifuge.ProtocolVersion1
+	if os.Getenv("CENTRIFUGE_PROTOCOL_V2") != "" {
+		log.Println("using client protocol v2")
+		protocolVersion = centrifuge.ProtocolVersion2
+	}
+
 	http.Handle("/connection/websocket", centrifuge.NewWebsocketHandler(node, centrifuge.WebsocketConfig{
-		WriteTimeout: time.Second,
+		ProtocolVersion: protocolVersion,
+		WriteTimeout:    time.Second,
 	}))
 
 	go func() {