--- a/_examples/ws_benchmarks/benchmark_nhooyr/main.go
+++ b/_examples/ws_benchmarks/benchmark_nhooyr/main.go
@@ -84,6 +84,11 @@ func (t *customWebsocketTransport) DisabledPushFlags() uint64 {
 	return centrifuge.PushFlagDisconnect
 }
 
+// Version of protocol.
+func (t *customWebsocketTransport) Version() centrifuge.ProtocolVersion {
+	return centrifuge.ProtocolVersion1
+}
+
 // Write ...
 func (t *customWebsocketTransport) Write(message []byte) error {
 	select {