--- a/_examples/ws_benchmarks/benchmark_nhooyr/main.go
+++ b/_examples/ws_benchmarks/benchmark_nhooyr/main.go
@@ -85,7 +85,7 @@ func (t *customWebsocketTransport) DisabledPushFlags() uint64 {
 }
 
 // Version of protocol.
-func (t *customWebsocketTransport) Version() centrifuge.ProtocolVersion {
+func (t *customWebsocketTransport) ProtocolVersion() centrifuge.ProtocolVersion {
 	return centrifuge.ProtocolVersion1
 }